Fixed bug; error handling
Moved MouseReleaseEvent handler to the card widget from the review
widget. Handler being on the review widget meant that the clicks on
buttons were also catched by it which resulted in undesirable flipping
of the next flashcard.

If some card is missing during a review the review is now aborted, and the home
screen is displayed accompanied by an error message.